Bloc FTW!
Added my own implementation of ViewModel and ViewModelProvider,
Added the ViewModelProvider directly on the navigation,
that way the views have no concern of where they're coming from.
Also, because was easy, that was a generic object provider.
I later found a package named provider that does the same,
but it was nice for learning.